test(reactive): move externals tests to new file
